requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.3.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
greenlet==1.1.2
humanize==4.0.0
kombu==5.2.4
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
